requirements.txt
playwright
python-dotenv
pytesseract
opencv-python-headless
numpy